Fix UI Changes and Design
Description

This pull request adds a new feature to the app that allows users to upload multiple images at once. The feature is implemented in the PhotosComponent and it includes validation for file type and size.

Changes

Added handleFileInput function to handle file input and validation.
Added deleteFile function to allow users to delete selected files.
Modified renderPhotos function to display the selected images.
Added state variables selectedFiles, errorMessage, and isFilePicked to manage the state of the component.
Modified the layout and styling of the component to display the selected images in a grid.
Added a new icon for the upload button.
